backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.34.0
python-multipart==0.0.21
python-dotenv==1.0.1

langchain-text-splitters==1.0.0
langchain-ollama==1.0.1
langchain-chroma==1.1.0
sentence-transformers==3.0.1

chromadb==1.4.0

pypdf==6.5.0
onnxruntime==1.23.2
pymupdf
python-magic
langfuse==2.51.4
